+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * add `apropos` function
 * new incremental Lexer that allow to modify parser while it's running
 * add SRFI-10 sharp-comma
+* new syntax try..catch..finally
 ### Bugfix
 * fix resolving promises on vectors and objects
 * fix context in methods of plain objects [#103](https://github.com/jcubic/lips/issues/103)
@@ -21,6 +22,7 @@
 * fix calling port methods (e.g. calling `get-output-string` function)
 * fix prism highlighting of multiple regular expression
 * fix parsing chars for Unicode outside of BMP
+* fix when try..catch is promise that rejects and catch throws
 
 ## 1.0.0-beta.10
 ### Breaking